[Oztechan/CCC#1931] Move widget viewmodel to android
