+++ android/ui/widget/android-ui-widget.gradle.kts
@@ -39,7 +39,7 @@ dependencies {
         }
     }
 
-    implementation(project(Modules.Client.ViewModel.widget))
+    implementation(project(Modules.Android.ViewModel.widget))
 
     implementation(project(Modules.Common.Core.model))
