+++ android/ui/widget/src/main/kotlin/com/oztechan/ccc/android/ui/widget/AppWidget.kt
@@ -3,7 +3,7 @@ package com.oztechan.ccc.android.ui.widget
 import androidx.compose.runtime.Composable
 import androidx.glance.appwidget.GlanceAppWidget
 import com.oztechan.ccc.android.ui.widget.content.WidgetView
-import com.oztechan.ccc.client.viewmodel.widget.WidgetViewModel
+import com.oztechan.ccc.android.viewmodel.widget.WidgetViewModel
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
