--- client/viewmodel/widget/client-viewmodel-widget.gradle.kts
@@ -1,100 +1,0 @@
-plugins {
-    @Suppress("DSL_SCOPE_VIOLATION")
-    libs.plugins.apply {
-        id(multiplatform.get().pluginId)
-        id(androidLib.get().pluginId)
-        alias(ksp)
-    }
-}
-kotlin {
-    android()
-
-    iosX64()
-    iosArm64()
-    iosSimulatorArm64()
-
-    @Suppress("UNUSED_VARIABLE")
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                libs.common.apply {
-                    implementation(koinCore)
-                }
-
-                Modules.Common.Core.apply {
-                    implementation(project(model))
-                }
-
-                Modules.Client.Core.apply {
-                    implementation(project(viewModel))
-                    implementation(project(shared))
-                }
-
-                Modules.Client.Storage.apply {
-                    implementation(project(app))
-                    implementation(project(calculation))
-                }
-
-                Modules.Client.DataSource.apply {
-                    implementation(project(currency))
-                }
-
-                Modules.Client.Service.apply {
-                    implementation(project(backend))
-                }
-            }
-        }
-        val commonTest by getting {
-            dependencies {
-                libs.common.apply {
-                    implementation(test)
-                    implementation(mockative)
-                    implementation(coroutinesTest)
-                }
-            }
-        }
-
-        val androidMain by getting
-        val androidTest by getting
-
-        val iosX64Main by getting
-        val iosArm64Main by getting
-        val iosSimulatorArm64Main by getting
-        val iosMain by creating {
-            dependsOn(commonMain)
-            iosX64Main.dependsOn(this)
-            iosArm64Main.dependsOn(this)
-            iosSimulatorArm64Main.dependsOn(this)
-        }
-        val iosX64Test by getting
-        val iosArm64Test by getting
-        val iosSimulatorArm64Test by getting
-        val iosTest by creating {
-            dependsOn(commonTest)
-            iosX64Test.dependsOn(this)
-            iosArm64Test.dependsOn(this)
-            iosSimulatorArm64Test.dependsOn(this)
-        }
-    }
-}
-
-dependencies {
-    ksp(libs.processors.mockative)
-}
-
-ksp {
-    arg("mockative.stubsUnitByDefault", "true")
-}
-
-@Suppress("UnstableApiUsage")
-android {
-    ProjectSettings.apply {
-        namespace = Modules.Client.ViewModel.widget.packageName
-        compileSdk = COMPILE_SDK_VERSION
-
-        defaultConfig {
-            minSdk = MIN_SDK_VERSION
-            targetSdk = TARGET_SDK_VERSION
-        }
-    }
-}
