--- client/viewmodel/widget/src/commonMain/kotlin/com/oztechan/ccc/client/viewmodel/widget/di/ClientViewModelWidgetModule.kt
@@ -1,9 +1,0 @@
-package com.oztechan.ccc.client.viewmodel.widget.di
-
-import com.oztechan.ccc.client.viewmodel.widget.WidgetViewModel
-import org.koin.dsl.module
-
-val clientViewModelWidgetModule = module {
-    // widget viewModel has to be singleton
-    single { WidgetViewModel(get(), get(), get(), get()) }
-}
